requirements.txt
streamlit
pydantic-ai-slim[tavily]
python-dotenv
